+++ calabash/Classes/CalabashServer.m
@@ -43,6 +43,8 @@
 #import "LPShakeRoute.h"
 #import "LPSuspendAppRoute.h"
 #import "LPReflectionRoute.h"
+#import "LPSetDeviceOrientationRoute.h"
+#import "LPStatusBarRoute.h"
 
 @interface CalabashServer ()
 - (void) start;
@@ -201,6 +203,14 @@
     LPReflectionRoute *reflectionRoute = [LPReflectionRoute new];
     [LPRouter addRoute:reflectionRoute forPath:@"reflection"];
     [reflectionRoute release];
+
+    LPSetDeviceOrientationRoute *orientationRoute = [LPSetDeviceOrientationRoute new];
+    [LPRouter addRoute:orientationRoute forPath:@"setDeviceOrientation"];
+    [orientationRoute release];
+
+    LPStatusBarRoute *statusBarRoute = [LPStatusBarRoute new];
+    [LPRouter addRoute:statusBarRoute forPath:@"statusBar"];
+    [statusBarRoute release];
 
     _httpServer = [[[LPHTTPServer alloc] init] retain];